Remove `eventstore.Result`.
