+++ persistence/subsystem/eventstore/repository.go
@@ -3,4 +3,3 @@ package eventstore
 import "github.com/dogmatiq/infix/persistence"
 
 type Repository = persistence.EventRepository
-type Result = persistence.EventResult
